requirements.txt
Django
Pillow
num2words
